+++ src/index.ts
@@ -9,8 +9,6 @@ import { WebSocket } from "ws";
 import { v4 as randomUUID } from "uuid";
 import { CloseCodes } from "./CloseCodes";
 
-let errorHappened = false;
-
 type errorCallbackType = (reason?: string) => void;
 
 type ZilaWSCallback = (...args: any[]) => any;
@@ -24,7 +22,7 @@ export enum WSStatus {
 
 interface WSMessage {
   identifier: string;
-  message: any[] | null;
+  message: any[] | any | null;
   callbackId: string | null;
 }
 
@@ -49,6 +47,20 @@ interface ICallableLocalEvents {
    * @returns
    */
   onRawMessageRecieved: (message: string) => void;
+
+  /**
+   * Runs when the server sends down one or multiple new cookies.
+   * @param cookieString
+   * @returns
+   */
+  onCookieSet: (cookieString: string) => void;
+
+  /**
+   * Runs when the server deletes a cookie
+   * @param cookieName
+   * @returns
+   */
+  onCookieDelete: (cookieName: string) => void;
 }
 
 export class ZilaConnection {
@@ -59,6 +71,7 @@ export class ZilaConnection {
   private connection: WebSocket | globalThis.WebSocket | undefined;
   private errorCallback: errorCallbackType | undefined;
   private _status: WSStatus = WSStatus.OPENING;
+  public maxWaiterTime: number = 1200;
 
   public get status(): WSStatus {
     return this._status;
@@ -93,13 +106,28 @@ export class ZilaConnection {
       let ws: WebSocket | globalThis.WebSocket | undefined;
 
       try {
+        //Can't test this with Node.
         /* istanbul ignore next */
         if (typeof window !== "undefined" && typeof window.document !== "undefined") {
           ws = new window.WebSocket(wsUrl);
+          ws.onerror = (ev) => {
+            console.error(JSON.stringify(ev));
+            errorCallback?.call(ev.type);
+            console.error("Disconnected from WebSocket server.");
+          };
         } else {
           ws = new WebSocket(wsUrl, {
             rejectUnauthorized: !allowSelfSignedCert,
+            headers: {
+              "s-type": 1,
+            },
           });
+
+          ws.onerror = (ev) => {
+            console.error(JSON.stringify(ev));
+            errorCallback?.call(ev.message);
+            console.error("Disconnected from WebSocket server.");
+          };
         }
       } catch (error) {
         const errorMessage = (error as Error).stack?.split("\n")[0];
@@ -124,14 +152,6 @@ export class ZilaConnection {
     this.errorCallback = errorCallback;
     this._status = WSStatus.OPENING;
 
-    this.connection.onerror = async () => {
-      this.status = WSStatus.ERROR;
-
-      this.errorCallback?.call(undefined);
-
-      console.error("Disconnected from WebSocket server.");
-    };
-
     this.connection.onopen = () => {
       this.status = WSStatus.OPEN;
     };
@@ -147,7 +167,7 @@ export class ZilaConnection {
       if (this.errorCallback) this.errorCallback(reason);
     };
 
-    this.connection.onmessage = (ev: any) => {
+    this.connection.onmessage = (ev: MessageEvent<any>) => {
       this.callEventHandler(ev.data.toString());
     };
   }
@@ -165,8 +185,31 @@ export class ZilaConnection {
 
     const msgObj: WSMessage = JSON.parse(msg);
 
-    if (!errorHappened) {
-      errorHappened = true;
+    /* istanbul ignore next */
+    if (msgObj.identifier[0] == "@") {
+      if (msgObj.identifier == "@SetCookie") {
+        if (typeof window !== "undefined" && typeof window.document !== "undefined") {
+          const cookieString = msgObj.message as unknown as string;
+          document.cookie = cookieString;
+
+          if (this.localEventCallbacks.onCookieSet)
+            for (const cb of this.localEventCallbacks.onCookieSet) {
+              cb(cookieString);
+            }
+        }
+      } else if (msgObj.identifier == "@DelCookie") {
+        if (typeof window !== "undefined" && typeof window.document !== "undefined") {
+          const cookieName = msgObj.message as unknown as string;
+          document.cookie = `${cookieName}=; expires=${new Date(0)}; path=/;`;
+
+          if (this.localEventCallbacks.onCookieDelete)
+            for (const cb of this.localEventCallbacks.onCookieDelete) {
+              cb(cookieName);
+            }
+        }
+      }
+
+      return;
     }
 
     if (this.localEventCallbacks.onMessageRecieved) {
@@ -186,50 +229,119 @@ export class ZilaConnection {
   }
 
   /**
-   * Calls an eventhandler on the serverside.
-   * @param {string} identifier The callback's name on the serverside.
-   * @param {any|undefined} data Arguments that shall be passed to the callback as parameters (optional)
+   * Returns a JSON serialized message object.
+   * @param identifier
+   * @param data
+   * @param callbackId
+   * @param isBuiltIn
+   * @returns
    */
-  public send(identifier: string, ...data: any[]): void {
-    if (typeof data == "function" || data.filter((el) => typeof el == "function").length > 0) {
-      throw new Error("Passing functions to the server is prohibited.");
-    }
-
-    const msg: WSMessage = {
-      callbackId: null,
+  private getMessageJSON(
+    identifier: string,
+    data: any[] | null,
+    callbackId: string | null,
+    isBuiltIn: boolean = false
+  ): string {
+    /* istanbul ignore next */
+    return JSON.stringify({
+      identifier: isBuiltIn ? identifier : "@" + identifier,
       message: data,
-      identifier: identifier,
-    };
-
-    this.connection!.send(JSON.stringify(msg));
+      callbackId: callbackId,
+    });
   }
 
   /**
-   * Calls an async callback on the serverside. Gets a value back from the serverside or just waits for the eventhandler to finish.
-   * @param {string} identifier The callback's name on the serverside.
+   * Send function for built-in systems
+   * @param {string} identifier The callback's name on the server-side.
+   * @param {any|undefined} data Arguments that shall be passed to the callback as parameters (optional)
+   */
+  /* istanbul ignore next */
+  private bSend(identifier: string, ...data: any[]) {
+    this.connection!.send(this.getMessageJSON(identifier, data, null, true));
+  }
+
+  /**
+   * Calls an eventhandler on the server-side.
+   * @param {string} identifier The callback's name on the server-side.
+   * @param {any|undefined} data Arguments that shall be passed to the callback as parameters (optional)
+   */
+  public send(identifier: string, ...data: any[]): void {
+    this.connection!.send(this.getMessageJSON(identifier, data, null));
+  }
+
+  /**
+   * Calls an async callback on the server-side. Gets a value back from the server-side or just waits for the eventhandler to finish.
+   * @param {string} identifier The callback's name on the server-side.
    * @param {any|undefined} data Arguments that shall be passed to the callback as parameters (optional)
    * @returns {Promise<any>}
    */
-  public async waiter(identifier: string, ...data: any[]): Promise<any> {
-    return new Promise((resolve) => {
-      if (typeof data == "function" || data.filter((el) => typeof el == "function").length > 0) {
-        throw new Error("Passing functions to the server is prohibited.");
-      }
-
+  public async waiter<T>(identifier: string, ...data: any[]): Promise<T | undefined> {
+    return new Promise(async (resolve) => {
       const uuid = randomUUID();
 
-      this.onceMessageHandler(uuid, (args: any): void => {
-        resolve(args);
-      });
+      let timeout: NodeJS.Timeout;
 
-      const msg: WSMessage = {
-        callbackId: uuid,
-        message: data,
-        identifier: identifier,
-      };
+      resolve(
+        Promise.any([
+          new Promise(r => {
+            this.setMessageHandler(uuid, (args: any[]): void => {
+              clearTimeout(timeout);
+              this.removeMessageHandler(uuid);
+              r(args);
+            });
+          }),
+          new Promise((_r) => {
+            timeout = setTimeout(() => {
+              _r(undefined);
+            }, this.maxWaiterTime);
+          })
+        ]) as Promise<T | undefined>
+      );
 
-      this.connection!.send(JSON.stringify(msg));
+      this.connection!.send(this.getMessageJSON(identifier, data, uuid));
     });
+  }
+
+  /**
+  * Calls an eventhandler on the server-side for the specified client. Gets a value of T type back from the client or just waits for the eventhandler to finish.
+  * @param {string} identifier The callback's name on the client-side.
+  * @param {number} maxWaitingTime The maximum time this waiter will wait for the client.
+  * @param {any|undefined} data Arguments that shall be passed to the callback as parameters (optional)
+  * @returns {Promise<T | undefined>}
+  */
+    public waiterTimeout<T>(identifier: string, maxWaitingTime: number, ...data: any[]): Promise<T | undefined> {
+      return new Promise(async (resolve) => {
+        const uuid = randomUUID();
+  
+        let timeout: NodeJS.Timeout;
+  
+        resolve(
+          Promise.any([
+            new Promise(r => {
+              this.setMessageHandler(uuid, (args: any[]): void => {
+                clearTimeout(timeout);
+                this.removeMessageHandler(uuid);
+                r(args);
+              });
+            }),
+            new Promise((_r) => {
+              timeout = setTimeout(() => {
+                _r(undefined);
+              }, maxWaitingTime);
+            })
+          ]) as Promise<T | undefined>
+        );
+  
+        this.connection!.send(this.getMessageJSON(identifier, data, uuid));
+      });
+    }
+
+  /**
+   * Sync cookies to the server-side.
+   */
+  /* istanbul ignore next */
+  public syncCookies() {
+    this.bSend("SyncCookies", document.cookie);
   }
 
   /**
@@ -237,7 +349,7 @@ export class ZilaConnection {
    *
    * EventListeners are made for local events only like for example: `onStatusChanged` that runs when the WebSocket client's connection status gets changed.
    *
-   * IF YOU WANT TO REGISTER A **MessageHandler**, THAT RUNS WHEN THE **SERVERSIDE** ASKS FOR IT, YOU **SHOULD** USE `registerMessageHandler`
+   * IF YOU WANT TO REGISTER A **MessageHandler**, THAT RUNS WHEN THE **server-side** ASKS FOR IT, YOU **SHOULD** USE `registerMessageHandler`
    * @param eventType
    * @param callback
    */
@@ -277,7 +389,7 @@ export class ZilaConnection {
   /**
    * Registers an EventListener that can be called only once.
    *
-   * If you want to register a MessageListener (what is called by serverside) only once you should use `onceMessageHandler`
+   * If you want to register a MessageListener (what is called by server-side) only once you should use `onceMessageHandler`
    * @param eventType
    * @param callback
    */
@@ -293,7 +405,7 @@ export class ZilaConnection {
   }
 
   /**
-   * Registers an eventhandler on the clientside that'll run when the server asks for it.
+   * Registers an eventhandler on the client-side that'll run when the server asks for it.
    * Can get overrided with using the same identifier.
    * @param identifier The eventhandler's name. **Must not start with `[ZilaWS]:`!**
    * @param callback The eventhandler.
@@ -303,7 +415,7 @@ export class ZilaConnection {
   }
 
   /**
-   * Removes a clientside MessageHandler.
+   * Removes a client-side MessageHandler.
    * @param identifier
    */
   public removeMessageHandler(identifier: string): void {
@@ -318,10 +430,7 @@ export class ZilaConnection {
   public onceMessageHandler(identifier: string, callback: ZilaWSCallback): void {
     this.callbacks[identifier] = async (...args: any[]) => {
       this.removeMessageHandler(identifier);
-      //const ret = callback(...args);
-
       const ret = await Promise.resolve(callback(...args));
-
       return ret;
     };
   }
@@ -383,8 +492,3 @@ export class ZilaConnection {
 export async function connectTo(wsUrl: string, errorCallback?: errorCallbackType): Promise<ZilaConnection> {
   return ZilaConnection.connectTo(wsUrl, errorCallback);
 }
-// const exporter = {
-//   connectTo: connectTo,
-//   WSStatus: WSStatus,
-//   ZilaConnection: ZilaConnection,
-// };